requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
bidict==0.22.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.1
Flask==2.1.3
Flask-SocketIO==5.2.0
flatbuffers>=2.0
frozenlist==1.3.1
future==0.18.2
gast==0.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gunicorn==20.1.0
grpcio==1.46.3
h5py==3.7.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
protobuf==3.19.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.1
pyu2f==0.1.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.0
Send2Trash==1.8.0
setuptools==63.4.1
six==1.15.0
tensorboard==2.9.1
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0rc0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.1
wheel==0.37.1
win-inet-pton==1.1.0
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.1
